pyproject.toml
[project]
name = "jain-ai"
version = "0.1.0"
description = "Modular Flask AI assistant with RAG, OCR, and session-based chat."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "groq>=0.18.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "pillow>=10.0.0",
    "pytest>=8.0.0",
]

[tool.uv]
package = false
